Transition and merge code, remove unused imports, fixed incorrect import pathing
