+++ Develop/client/package.json
@@ -1,7 +1,7 @@
 {
   "name": "zeropath-client",
   "private": true,
-  "version": "0.0.0",
+  "version": "1.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@apollo/client": "^3.11.8",
+    "@tailwindcss/forms": "^0.5.9",
     "bootstrap": "^5.2.3",
     "jwt-decode": "^3.1.2",
     "react": "^18.2.0",
@@ -27,6 +28,8 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.3.4",
+    "postcss": "^8.4.47",
+    "tailwindcss": "^3.4.13",
     "vite": "^5.1.6"
   }
 }
