+++ Develop/client/postcss.config.cjs
@@ -1,3 +1,4 @@
+/* eslint-disable no-undef */
 module.exports = {
   plugins: {
     tailwindcss: {},
