+++ Develop/client/src/App.jsx
@@ -1,40 +1,10 @@
-import "./App.css";
-import Navbar from "./components/Navbar";
-import { Outlet } from "react-router-dom";
-
-import {
-  ApolloClient,
-  InMemoryCache,
-  ApolloProvider,
-  createHttpLink,
-} from "@apollo/client";
-import { setContext } from "@apollo/client/link/context";
-
-const httpLink = createHttpLink({
-  uri: "/graphql",
-});
-
-const authLink = setContext((_, { headers }) => {
-  const token = localStorage.getItem("id_token");
-  return {
-    headers: {
-      ...headers,
-      authorization: token ? `Bearer ${token}` : "",
-    },
-  };
-});
-
-const client = new ApolloClient({
-  link: authLink.concat(httpLink),
-  cache: new InMemoryCache(),
-});
+import LandingPage from "./pages/LandingPage/LandingPage";
 
 function App() {
   return (
-    <ApolloProvider client={client}>
-      <Navbar />
-      <Outlet />
-    </ApolloProvider>
+    <div className="App">
+      <LandingPage />
+    </div>
   );
 }
