+++ Develop/client/src/components/Canvas/Canvas.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import useCanvas from "../../hooks/useCanvas";
 
 const Canvas = () => {
