+++ Develop/client/src/main.jsx
@@ -1,29 +1,11 @@
+import React from "react";
 import ReactDOM from "react-dom/client";
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import "bootstrap/dist/css/bootstrap.min.css";
+import "./styles/index.css";
+import App from "../src/App";
 
-import App from "./App.jsx";
-import SearchBooks from "./pages/SearchBooks.jsx";
-import SavedBooks from "./pages/SavedBooks.jsx";
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <App />,
-    errorElement: <h1 className="display-2">Wrong page!</h1>,
-    children: [
-      {
-        index: true,
-        element: <SearchBooks />,
-      },
-      {
-        path: "/saved",
-        element: <SavedBooks />,
-      },
-    ],
-  },
-]);
-
-ReactDOM.createRoot(document.getElementById("root")).render(
-  <RouterProvider router={router} />
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
 );
