+++ Develop/client/src/utils/animations.jsx
