+++ Develop/package.json
@@ -14,6 +14,7 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
+    "autoprefixer": "^10.4.20",
     "concurrently": "^9.0.1"
   },
   "dependencies": {
