--- OldCode/src/App.js
@@ -1,12 +1,0 @@
-import React from "react";
-import LandingPage from "./components/LandingPage/LandingPage";
-
-function App() {
-  return (
-    <div className="App">
-      <LandingPage />
-    </div>
-  );
-}
-
-export default App;
